Users

+++ model/src/main/java/ar/edu/itba/paw/model/User.java
@@ -1,12 +1,12 @@
 package ar.edu.itba.paw.model;
 
 public class User {
-    private long id;
+    private int id;
     private String username;
     private String password;
     private String mail;
 
-    public User(long id, String username, String password, String mail) {
+    public User(int id, String username, String password, String mail) {
         this.id = id;
         this.username = username;
         this.password = password;
@@ -21,7 +21,7 @@ public class User {
         return password;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
